RISC-V: Use Relative for run_jump_pc return type
